pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "google-genai>=1.46.0",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "plotly>=6.3.1",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
    "streamlit-camera-input-live>=0.2.0",
]
